requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
websockets==12.0
python-multipart==0.0.6
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
aiofiles==23.2.1
cryptography==45.0.6
gpsdclient==1.4.5
pillow==10.1.0